RestPractice/requirements.txt
Django>=4.1,<5
djangorestframework

